Patch against Synergy Virus v2.9.0
Fixed an issue where the game was frozen when all purchase upgrades reached 2 ^ 53.
Accelerators and Multipliers and Accelerator Boost, which were capped to 1E15 were canceled.
The number of architects owned has been fixed to 0 if it exceeds 2 ^ 50. This was a significant reduction in COIN and DIAMOND and Mythos and PARTICLE production.
Offering and obtainium will freeze games if you exceed 2 ^ 1024. They can cap to 1E300. This has already occurred in some players.
Caps at 1E300 for all ANT SACRIFICE rewards. It was the cause of frozen the game.
T-challenge increased 9001 completion to 1337E40x. Auto Challenge Sweep [ON] does not work.
If Artemis' exceeds 2 ^ 1024, the problem of galactic crumbs would be 0 in 1E300 cap. This occurs in an environment where the opened cube exceeds E100.
Ant Multiplier could be calculated up to about E600. Ants will work until Player.AndsActrificepoints reaches 2 ^ 1024.
From power notation, we added to EE1E99 such as SXDC (SEXDECILLION). We adopted VG from Vigintillion.
Merged a small number if it is not a power notation. This was mainly seen in coin architecture.
G [GlobalCoinMultiplier '] has been fixed in part of calculation of 2 ^ 1024 in part of the calculation of [GlobalCoinMultiplier']. It occurred from the coin EE + 130.
G ['TotalMultiPlierboost'] has been fixed in the 1E300 cap that there was a case that more than 2 ^ 1024.
Fixed that Crystal Shards could be negative. This was the cause of Offerings to become NaN.

Number changed to DECIMAL only has a problem with problems. The remainder should work by the developer.
All caps added in a place that could not be handled is 1E300.
C15 EXPONENT E24 or higher game balance changes. This is because a synergy stopped with a defect works again.
C15 EXPONENT E50 or higher tests are insufficient, but we confirmed that the game works with EE242 coin (1E-300 TAX ​​test).

This patch was made to test the defect. Operation can not be guaranteed.